LPS-58897 Move out to bundle
