+++ modules/portal-cache/portal-cache/src/com/liferay/portal/cache/AggregatedPortalCacheListener.java
@@ -12,7 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.kernel.cache;
+package com.liferay.portal.cache;
+
+import com.liferay.portal.kernel.cache.PortalCache;
+import com.liferay.portal.kernel.cache.PortalCacheException;
+import com.liferay.portal.kernel.cache.PortalCacheListener;
+import com.liferay.portal.kernel.cache.PortalCacheListenerScope;
+import com.liferay.portal.kernel.cache.SkipReplicationThreadLocal;
 
 import java.io.Serializable;
