+++ modules/portal-cache/portal-cache/src/com/liferay/portal/cache/PortalCacheWrapper.java
@@ -12,7 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.kernel.cache;
+package com.liferay.portal.cache;
+
+import com.liferay.portal.kernel.cache.PortalCache;
+import com.liferay.portal.kernel.cache.PortalCacheListener;
+import com.liferay.portal.kernel.cache.PortalCacheListenerScope;
+import com.liferay.portal.kernel.cache.PortalCacheManager;
 
 import java.io.Serializable;
